pyproject.toml
[project]
name = "z-image-turbo"
version = "0.1.0"
description = "Fast image generation with Z-Image-Turbo, featuring a terminal viewer with Kitty graphics protocol and Gradio web UI. Optimized for Apple Silicon and CUDA."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
dependencies = [
    "accelerate>=1.12.0",
    "diffusers @ git+https://github.com/huggingface/diffusers",
    "gradio>=5.0.0",
    "pillow>=12.0.0",
    "requests>=2.31.0",
    "torch>=2.9.1",
    "torchao>=0.7.0",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

